Use `uname` for some substitutions
With this change, we use `getauxval` to get the platform on Linux when
that function is available, and when it isn't, we use the POSIX `uname`
to retrieve that info. The `utsname` struct gets passed around as well
so that substitutions other than `$PLATFORM` can be applied downstream
with only one additional function parameter.